requirements.txt
cryptography>=41.0.0
pyyaml>=6.0
regex>=2023.0.0
numpy>=1.24.0

# Optional — only if using Redis or encrypted_db backends
# redis>=5.0.0
# sqlalchemy>=2.0.0
